Update camping-bel-air-no-user-icon.js
${campingName}

The error TypeError: Cannot read properties of undefined (reading 'CAMPING_NAME') at line 25 of your JavaScript code occurs because you are trying to access CampingChatbot.config.CAMPING_NAME within the definition of CampingChatbot.config itself. During the static initialization of the class, when the JavaScript engine is evaluating the config object, it seems that CampingChatbot.config is not yet fully assigned, leading to it being undefined at the moment the header property of the COPY object is being initialized.